# crypto: use btcec/v2 for no-cgo