This updates the no-cgo code in the `crypto` package to use the new github.com/btcsuite/btcd/btcec/**v2** module instead of the older btcec package that was part of the main github.com/btcsuite/btcd module.

This is nice as it removes the dependency on the legacy `btcd` mega-module that does not yet follow semantic import versioning (has breaking Go API changes between versions like v0.21 and v0.22) with the [new `btcec/v2` module](https://github.com/btcsuite/btcd/releases/tag/btcec%2Fv2.1.2) that does follow SIV, plus it brings performance improvements:

```
name                   old time/op  new time/op  delta
EcrecoverSignature-32   198µs ± 0%   144µs ± 0%  -27.11%
VerifySignature-32      177µs ± 0%   128µs ± 0%  -27.44%
DecompressPubkey-32    20.9µs ± 0%  10.1µs ± 0%  -51.51%
```

The tests that passed previously with `CGO_ENABLED=0` still pass, and the secp256k1 fuzzing results show no discrepancies with the cgo version.

tests/fuzzers/secp256k1:
```
$ CGO_ENABLED=0 go-fuzz
2022/03/12 17:58:32 workers: 16, corpus: 60 (3s ago), crashers: 0, restarts: 1/0, execs: 0 (0/sec), cover: 0, uptime: 3s
2022/03/12 17:58:35 workers: 16, corpus: 60 (6s ago), crashers: 0, restarts: 1/0, execs: 0 (0/sec), cover: 949, uptime: 6s
2022/03/12 17:58:38 workers: 16, corpus: 60 (9s ago), crashers: 0, restarts: 1/4357, execs: 209144 (23236/sec), cover: 949, uptime: 9s
...
2022/03/12 17:59:44 workers: 16, corpus: 61 (17s ago), crashers: 0, restarts: 1/9514, execs: 4938271 (65843/sec), cover: 949, uptime: 1m15s
2022/03/12 17:59:47 workers: 16, corpus: 61 (20s ago), crashers: 0, restarts: 1/9592, execs: 5151037 (66038/sec), cover: 949, uptime: 1m18s
```

The most notable recent breaking change in the main `btcd` module that was previously required is that with the introduction of the `btcec/v2` module, there is no longer a btcec package in the main `btcd` module, which creates build errors when the latest revisions of btcd are used by either go-ethereum or transitive dependencies.  Moving to the `btcec/v2` module resolves this issue.

CC @fjl