@chappjc 2022-03-14
Pushed some minor changes to make it a bit more defensive: checking expected lengths, using package-level consts, and zeroing the temporary `btcec.PrivateKey` constructed in `Sign`.  diff: https://github.com/ethereum/go-ethereum/compare/8e1026cab147b84a43ab69d03abadf1df7c9e41a..2f96c2bf0a175ca9b893b1073aa81961b4dd5014